Stricter type checking between Python and Pythran result
- ceil/floor/trunc return an int
- max/min don't do type promotion
- exception's args can be stored as dynamic_tuple instead of list
- rank returns an int and not a size_t
- normalize_compare generates bool and not int

This fixes #339 that actually got fixed a while ago, but now we are sure
it won't regress.

Still allow int64 <> long, bool_ <> bool float32 <> float conversion.

+++ pythran/pythonic/include/numpy/rank.hpp
@@ -9,7 +9,7 @@ PYTHONIC_NS_BEGIN
 namespace numpy
 {
   template <class E>
-  size_t rank(E const &expr);
+  long rank(E const &expr);
 
   DEFINE_FUNCTOR(pythonic::numpy, rank);
 }
